Simplify sending emails for action
Apply consistent approach across email workflow rules by status and by action. Future planning and implementation may be required to ensure transactional nature of applicable endpoints and a mechanism to ensure emails are sent and retried if any glitches occur.
